requirements.txt
yfinance
psycopg2-binary
pandas_market_calendars
pandas
numpy
sqlalchemy
dotenv